backend/lambda/agent-wireframe/requirements.txt
boto3==1.34.0
pydantic==2.5.0
openai>=1.12.0
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
httpx>=0.25.0

